Move the various operation classes to their own files.
