+++ Evaluators/AggregateOp.cs
@@ -1,0 +1,52 @@
+
+/*
+ *  MetaphysicsIndustries.Solus
+ *  Copyright (C) 2006-2021 Metaphysics Industries, Inc., Richard Sartor
+ *
+ *  This library is free software; you can redistribute it and/or
+ *  modify it under the terms of the GNU Lesser General Public
+ *  License as published by the Free Software Foundation; either
+ *  version 3 of the License, or (at your option) any later version.
+ *
+ *  This library is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ *  Lesser General Public License for more details.
+ *
+ *  You should have received a copy of the GNU Lesser General Public
+ *  License along with this library; if not, write to the Free Software
+ *  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301
+ *  USA
+ *
+ */
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using MetaphysicsIndustries.Solus.Exceptions;
+using MetaphysicsIndustries.Solus.Expressions;
+using MetaphysicsIndustries.Solus.Functions;
+using MetaphysicsIndustries.Solus.Macros;
+using MetaphysicsIndustries.Solus.Transformers;
+using MetaphysicsIndustries.Solus.Values;
+
+namespace MetaphysicsIndustries.Solus.Evaluators
+{
+    public abstract class AggregateOp
+    {
+        public abstract void Operate(IMathObject input,
+            SolusEnvironment env, BasicEvaluator evaluator);
+    }
+
+    public class AggregateOp<TIn, TOut> : AggregateOp
+    {
+        public TOut State;
+        public Func<TIn, TOut, TOut> Function;
+
+        public override void Operate(IMathObject input,
+            SolusEnvironment env, BasicEvaluator evaluator)
+        {
+            State = Function((TIn)input, State);
+        }
+    }
+}
