+++ Evaluators/FunctionAggregateOp.cs
@@ -1,0 +1,49 @@
+
+/*
+ *  MetaphysicsIndustries.Solus
+ *  Copyright (C) 2006-2021 Metaphysics Industries, Inc., Richard Sartor
+ *
+ *  This library is free software; you can redistribute it and/or
+ *  modify it under the terms of the GNU Lesser General Public
+ *  License as published by the Free Software Foundation; either
+ *  version 3 of the License, or (at your option) any later version.
+ *
+ *  This library is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ *  Lesser General Public License for more details.
+ *
+ *  You should have received a copy of the GNU Lesser General Public
+ *  License along with this library; if not, write to the Free Software
+ *  Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301
+ *  USA
+ *
+ */
+
+using MetaphysicsIndustries.Solus.Functions;
+
+namespace MetaphysicsIndustries.Solus.Evaluators
+{
+    public class FunctionAggregateOp : AggregateOp
+    {
+        public FunctionAggregateOp(Function function,
+            IMathObject initialState)
+        {
+            Function = function;
+            State = initialState;
+        }
+
+        public IMathObject State;
+        public readonly Function Function;
+
+        private readonly IMathObject[] _args = new IMathObject[2];
+        public override void Operate(IMathObject input,
+            SolusEnvironment env, BasicEvaluator evaluator)
+        {
+            _args[0] = input;
+            _args[1] = State;
+            var result = evaluator.Call(Function, _args, env);
+            State = result;
+        }
+    }
+}
